[not verified] refact -> add usePost hook
