+++ projects/packages/wp-admin-plus/src/components/post-status-label/index.js
@@ -1,32 +1,11 @@
 /**
  * External dependencies
  */
-import { useSelect } from '@wordpress/data';
-import { store as coreStore } from '@wordpress/core-data';
 import { Fragment } from '@wordpress/element';
+import usePost from '../../hooks/use-post';
 
 export default function PostStatusLabel( { id, postIds, status, type, statuses, fallbackText } ) {
-	const post = useSelect(
-		select => {
-			const posts = select( coreStore ).getEntityRecords( 'postType', type, {
-				include: postIds,
-				status,
-			} );
-
-			if ( ! posts?.length ) {
-				return;
-			}
-
-			const filteredPostById = posts.filter( item => item?.id === id );
-			if ( ! filteredPostById?.length ) {
-				return;
-			}
-
-			return filteredPostById[ 0 ];
-		},
-		[ postIds, id ]
-	);
-
+	const post = usePost( { id, postIds, type } );
 	if ( ! post?.status ) {
 		return fallbackText;
 	}
